LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/track/track-element/track-webvtt-align-text-line-position-expected.txt



FAIL Cue alignment, line and text position from settings assert_equals: expected (string) "auto" but got (number) -1
